Переименование всех значений, связанных с компонентами.
